requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
django==4.2.28
djangorestframework==3.16.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
python-decouple==3.8
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
